pyproject.toml
[project]
name = "eec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.8",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.1",
]
